phase-1/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.23
psycopg[binary]>=3.1.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.0
anthropic>=0.33.0
python-slugify>=8.0.1
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
